# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Apache Airavata HTTP API
  version: "1.0.0"
  description: |
    The REST API served by `airavata-server` (see `internal/server/server.go`).

    All request and response bodies are JSON (`Content-Type: application/json`).

    ## Authentication

    Authentication is a bearer token validated by CILogon introspection, carried as
    `Authorization: Bearer <token>`. A request without a token is anonymous rather
    than rejected at the door: each operation decides for itself whether an anonymous
    caller is acceptable. A token that is present but invalid is always `401`, with
    `WWW-Authenticate: Bearer error="invalid_token"`. If the identity provider cannot
    be reached the call fails `502`, because that is not the caller's fault.

    ## Authorization

    Two schemes are in play, depending on the resource:

    - **Platform roles** (`ADMIN`, `SUPER_ADMIN`) guard the shared catalogue —
      Slurm clusters, partitions, application templates and batch deployments.
      Reads of the catalogue are open to anyone, including anonymous callers;
      writes require a role.
    - **Ownership and sharing** guards the per-user resources — SSH keys, groups,
      cluster configs, SCP data storages, data products and processes. Ownership
      comes from the token and is never carried in a request body, so nothing can be
      created on someone else's behalf or handed over by editing it. None of these
      are readable anonymously.

...

...

...

...
    state.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/processes/{processId}:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [Processes]
      summary: Read a process
      description: Owner or admin. Every section the process carries is nested inside it.
      operationId: getProcess
      responses:
        "200":
          description: The process.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [Processes]
      summary: Update a process
      description: |
        Requires `ADMIN` or `SUPER_ADMIN` authority. The input and output mapping sets
        are replaced wholesale.
      operationId: updateProcess
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcessRequest" }
      responses:
        "200":
          description: The updated process.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProcessResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Processes]
      summary: Delete a process
      description: Requires `ADMIN` or `SUPER_ADMIN` authority.
      operationId: deleteProcess
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/processes/{processId}/data-staging-tasks:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [Processes]
      summary: List a process's data staging tasks
      operationId: listDataStagingTasks
      responses:
        "200":
          description: The process's staging tasks.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DataStagingTaskResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Processes]
      summary: Create a data staging task
      description: |
        The task belongs to the process in the path; there is no `processId` in the
        body, so a task cannot be moved to a different process by editing it.
      operationId: createDataStagingTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DataStagingTaskRequest" }
      responses:
        "201":
          description: The created task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DataStagingTaskResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/processes/{processId}/job-submission-tasks:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [Processes]
      summary: List a process's job submission tasks
      operationId: listJobSubmissionTasks
      responses:
        "200":
          description: The process's submission tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  { $ref: "#/components/schemas/JobSubmissionTaskResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Processes]
      summary: Create a job submission task
      operationId: createJobSubmissionTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobSubmissionTaskRequest" }
      responses:
        "201":
          description: The created task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobSubmissionTaskResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/processes/{processId}/job-monitoring-tasks:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [Processes]
      summary: List a process's job monitoring tasks
      operationId: listJobMonitoringTasks
      responses:
        "200":
          description: The process's monitoring tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  { $ref: "#/components/schemas/JobMonitoringTaskResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Processes]
      summary: Create a job monitoring task
      operationId: createJobMonitoringTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobMonitoringTaskRequest" }
      responses:
        "201":
          description: The created task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobMonitoringTaskResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/v1/processes/{processId}/interactive-command-tasks:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [Processes]
      summary: List a process's interactive command tasks
      operationId: listInteractiveCommandTasks
      responses:
        "200":
          description: The process's command tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  {
                    $ref: "#/components/schemas/InteractiveCommandTaskResponse",
                  }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Processes]
      summary: Create an interactive command task
      operationId: createInteractiveCommandTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              { $ref: "#/components/schemas/InteractiveCommandTaskRequest" }
      responses:
        "201":
          description: The created task.
          content:
            application/json:
              schema:
                { $ref: "#/components/schemas/InteractiveCommandTaskResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A CILogon access token, validated by RFC 7662 introspection. The bootstrap
        root token is also accepted; see INSTALL.md.
  parameters:
...
    ProcessId:
      name: processId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NoContent:
      description: Deleted. No body.
    BadRequest:
      description: |
        Malformed JSON, or a body that failed its constraints. Validation failures
        carry `fieldErrors`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            validation:
              summary: Validation failed
              value:
                status: 400
                error: Bad Request
                message: Validation failed
                fieldErrors:
                  - field: clusterName
                    message: Cluster name cannot be blank
    Unauthorized:
      description: |
        No token where one is required, or a token that is present but unusable. An
        invalid token also carries `WWW-Authenticate: Bearer error="invalid_token"`.
      headers:
        WWW-Authenticate:
          description: Present when the token was rejected.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated, but lacking the authority for this call.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such record — or, for a group, a record the caller has no standing to know
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            status: 404
            error: Not Found
            message: "Slurm cluster not found: c1d2e3f4-5678-4abc-9def-0123456789ab"
    Conflict:
      description: |
        The request collides with existing state — deleting an SSH key still in use,
        or a template that still has deployments.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadGateway:
      description: |
        The identity provider could not be reached to validate the token. Not the
        caller's fault, so it is not reported as an auth failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: |
        An unexpected server failure. The message is always `"Internal server error"`;
        the detail is logged rather than returned, since it may name internal state.

        A `502` is also possible on any authenticated call if the identity provider is
        unreachable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    FieldError:
      type: object
      description: One failed validation constraint.
      properties:
        field:
          type: string
          description: |
            The JSON path of the offending field, including indexes and nesting.
          examples:
            ["inputs[0].inputType", "defaultBatchJobConfig.wallTimeMinutes"]
        message: { type: string }
      required: [field, message]
    Error:
      type: object
      description: The envelope every failure response uses.
      properties:
        status:
          type: integer
          description: The HTTP status, repeated in the body.
        error:
          type: string
          description: The standard reason phrase for `status`.
        message:
          type: string
          description: Human-readable detail.
        fieldErrors:
          type: array
          description: |
            Present only on request-body validation failures, listing every
            constraint that failed so a client can report all of them at once.
          items: { $ref: "#/components/schemas/FieldError" }
      required: [status, error, message]

    # ----------------------------------------------------------- Enums
...
    DataStorageType:
      type: string
      enum: [SCP, HPC]
...
    ProcessType:
      type: string
      enum: [BATCH_JOB, CLOUD_JOB]
...
    OnFailureAction:
      type: string
      enum: [RETRY, SKIP, EXIT]
    BatchJobStatusType:
      type: string
      description: |
        Airavata's own states (`SUBMITTED`, `SUBMISSION_FAILED`, `CANCELLING`,
        `CANCELLED`) plus the Slurm mail states parsed from job notifications.
      enum:
        - SUBMITTED
        - SUBMISSION_FAILED
        - CANCELLING
        - CANCELLED
        - BEGIN
        - REQUEUE
        - INVALID_DEPEND
        - STAGE_OUT
        - TIME_LIMIT
        - TIME_LIMIT_90
        - TIME_LIMIT_80
        - TIME_LIMIT_50
        - FAIL
        - END

    # ------------------------------------------------------------ IAM
...
    BatchJobConfigRequest:
      type: object
      description: The resource request carried by a deployment or a run.
      properties:
        wallTimeMinutes: { type: integer, format: int64, exclusiveMinimum: 0 }
        allocation: { type: string }
        cpus: { type: [integer, "null"], format: int32 }
        mem: { type: [string, "null"] }
        memPerCpu: { type: [string, "null"] }
        ntasksPerNode: { type: [integer, "null"], format: int32 }
        cpusPerTask: { type: [integer, "null"], format: int32 }
        nodes: { type: [integer, "null"], format: int32 }
        ntasks: { type: [integer, "null"], format: int32 }
        gres: { type: [string, "null"] }
        gpus: { type: [integer, "null"], format: int32 }
        memPerGpu: { type: [string, "null"] }
        cpusPerGpu: { type: [string, "null"] }
        gpusPerNode: { type: [integer, "null"], format: int32 }
        constraints: { type: [string, "null"] }
      required: [wallTimeMinutes, allocation]
    BatchJobConfigResponse:
      type: object
      properties:
        batchJobConfigId: { type: string, format: uuid }
        wallTimeMinutes: { type: integer, format: int64 }
        allocation: { type: string }
        cpus: { type: [integer, "null"], format: int32 }
        mem: { type: [string, "null"] }
        memPerCpu: { type: [string, "null"] }
        ntasksPerNode: { type: [integer, "null"], format: int32 }
        cpusPerTask: { type: [integer, "null"], format: int32 }
        nodes: { type: [integer, "null"], format: int32 }
        ntasks: { type: [integer, "null"], format: int32 }
        gres: { type: [string, "null"] }
        gpus: { type: [integer, "null"], format: int32 }
        memPerGpu: { type: [string, "null"] }
        cpusPerGpu: { type: [string, "null"] }
        gpusPerNode: { type: [integer, "null"], format: int32 }
        constraints: { type: [string, "null"] }
      required: [batchJobConfigId, wallTimeMinutes, allocation]
...
    InputMapping:
      type: object
      description: |
        Binds one of the deployment template's declared inputs to a value for this
        run. `templateInputMappingId` is echoed back on reads but ignored on writes:
        an update replaces the mapping set wholesale.
      properties:
        templateInputMappingId: { type: string, format: uuid }
        templateInputId: { type: string, format: uuid }
        value:
          type: [string, "null"]
          description: |
            A JSON document, either `{"value": "..."}` for a single value or
            `{"values": [...]}` for a list.
      required: [templateInputId]
    OutputMapping:
      type: object
      properties:
        templateOutputMappingId: { type: string, format: uuid }
        templateOutputId: { type: string, format: uuid }
        value:
          type: [string, "null"]
          description: |
            A JSON document, either `{"value": "..."}` for a single value or
            `{"values": [...]}` for a list.
      required: [templateOutputId]
    BatchJobStatusResponse:
      type: object
      properties:
        batchProcessStatusId: { type: string, format: uuid }
        batchProcessId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BatchJobStatusType" }
        updatedAt:
          type: integer
          format: int64
          description: When the state was recorded, in milliseconds since the epoch.
      required: [batchProcessStatusId, batchProcessId, status, updatedAt]
    BatchProcessRequest:
      type: object
      description: |
        The `batchProcess` section of a process request. The resource request is
        carried here rather than copied from the deployment's default, which is what
        lets a caller ask for different resources for a particular run.
      properties:
        deploymentId: { type: string, format: uuid }
        jobId:
          type: [string, "null"]
          description: |
            The scheduler's identifier for the submitted job — writable rather than
            server-generated, because it is learned at submission time and recorded
            afterwards.
        batchJobConfig: { $ref: "#/components/schemas/BatchJobConfigRequest" }
        slurmClusterConfigId:
          type: string
          format: uuid
          description: |
            The cluster login config this run submits under. Required: a deployment
            names the machine but nobody's identity on it. Only a config the caller
            owns or that is shared with them is accepted.
        baseWorkDir:
          type: [string, "null"]
          description: |
            The parent directory on the cluster this run works under; it gets a
            subdirectory of its own beneath it, named for the process id, which is
            where file inputs are staged to and file outputs staged from.
        inputMappings:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/InputMapping" }
        outputMappings:
          type: [array, "null"]
          items: { $ref: "#/components/schemas/OutputMapping" }
      required: [deploymentId, slurmClusterConfigId, batchJobConfig]
    BatchProcessResponse:
      type: object
      description: |
        The `batchProcess` section of a process response. The nested config is a
        snapshot of what this run actually asked for.
      properties:
        batchProcessId: { type: string, format: uuid }
        deploymentId: { type: [string, "null"], format: uuid }
        jobId: { type: [string, "null"] }
        batchJobConfig:
          oneOf:
            - { $ref: "#/components/schemas/BatchJobConfigResponse" }
            - { type: "null" }
        slurmClusterConfigId: { type: string, format: uuid }
        baseWorkDir: { type: [string, "null"] }
        inputMappings:
          type: array
          items: { $ref: "#/components/schemas/InputMapping" }
        outputMappings:
          type: array
          items: { $ref: "#/components/schemas/OutputMapping" }
        batchJobStatuses:
          type: array
          description: Oldest first.
          items: { $ref: "#/components/schemas/BatchJobStatusResponse" }
        latestBatchJobStatus:
          oneOf:
            - { $ref: "#/components/schemas/BatchJobStatusResponse" }
            - { type: "null" }
      required:
        - batchProcessId
        - slurmClusterConfigId
        - inputMappings
        - outputMappings
        - batchJobStatuses
    ProcessRequest:
      type: object
      description: |
        The create/update payload for a process. There is no owner field. Which
        sections a body may carry follows from its type: a `BATCH_JOB` without a
        `batchProcess` describes no job at all, and a `batchProcess` on any other kind
        is a caller's mistake rather than something to store and ignore.
      properties:
        processType: { $ref: "#/components/schemas/ProcessType" }
        batchProcess:
          description: Required when `processType` is `BATCH_JOB`; rejected otherwise.
          oneOf:
            - { $ref: "#/components/schemas/BatchProcessRequest" }
            - { type: "null" }
      required: [processType]
    ProcessResponse:
      type: object
      description: |
        The read model for a process, with every section it carries nested inside it.
      properties:
        processId: { type: string, format: uuid }
        userId:
          type: [string, "null"]
          description: The owner, taken from the token at creation.
        processType:
          oneOf:
            - { $ref: "#/components/schemas/ProcessType" }
            - { type: "null" }
        lastStatusId: { type: [string, "null"], format: uuid }
        batchProcess:
          oneOf:
            - { $ref: "#/components/schemas/BatchProcessResponse" }
            - { type: "null" }
      required: [processId]
...
    DataStagingTaskRequest:
      type: object
      description: |
        Both paths are required: a staging task that names neither where data comes
        from nor where it goes describes no transfer at all.
      properties:
        sourceDataStorageId: { type: [string, "null"], format: uuid }
        sourceDataStorageType:
          oneOf:
            - { $ref: "#/components/schemas/DataStorageType" }
            - { type: "null" }
        destinationDataStorageId: { type: [string, "null"], format: uuid }
        destinationDataStorageType:
          oneOf:
            - { $ref: "#/components/schemas/DataStorageType" }
            - { type: "null" }
        sourcePath: { type: string }
        destinationPath: { type: string }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"], minimum: 0 }
        taskOrder: { type: [integer, "null"], minimum: 0 }
      required: [sourcePath, destinationPath]
    DataStagingTaskResponse:
      type: object
      properties:
        taskId: { type: string, format: uuid }
        processId: { type: [string, "null"], format: uuid }
        sourceDataStorageId: { type: [string, "null"], format: uuid }
        sourceDataStorageType:
          oneOf:
            - { $ref: "#/components/schemas/DataStorageType" }
            - { type: "null" }
        destinationDataStorageId: { type: [string, "null"], format: uuid }
        destinationDataStorageType:
          oneOf:
            - { $ref: "#/components/schemas/DataStorageType" }
            - { type: "null" }
        sourcePath: { type: [string, "null"] }
        destinationPath: { type: [string, "null"] }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"] }
        taskOrder: { type: [integer, "null"] }
      required: [taskId]
    JobSubmissionTaskRequest:
      type: object
      properties:
        jobId:
          type: [string, "null"]
          description: |
            The scheduler's identifier for the submitted job — writable rather than
            server-generated.
        workingDir: { type: [string, "null"] }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"], minimum: 0 }
        taskOrder: { type: [integer, "null"], minimum: 0 }
    JobSubmissionTaskResponse:
      type: object
      properties:
        taskId: { type: string, format: uuid }
        processId: { type: [string, "null"], format: uuid }
        jobId: { type: [string, "null"] }
        workingDir: { type: [string, "null"] }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"] }
        taskOrder: { type: [integer, "null"] }
      required: [taskId]
    JobMonitoringTaskRequest:
      type: object
      properties:
        jobId: { type: [string, "null"] }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"], minimum: 0 }
        taskOrder: { type: [integer, "null"], minimum: 0 }
    JobMonitoringTaskResponse:
      type: object
      properties:
        taskId: { type: string, format: uuid }
        processId: { type: [string, "null"], format: uuid }
        jobId: { type: [string, "null"] }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"] }
        taskOrder: { type: [integer, "null"] }
      required: [taskId]
    InteractiveCommandTaskRequest:
      type: object
      properties:
        command: { type: string }
        output:
          type: [string, "null"]
          description: |
            The result of running the command — writable for the same reason `jobId`
            is, recorded once whatever executed it knows the answer.
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"], minimum: 0 }
        taskOrder: { type: [integer, "null"], minimum: 0 }
      required: [command]
    InteractiveCommandTaskResponse:
      type: object
      properties:
        taskId: { type: string, format: uuid }
        processId: { type: [string, "null"], format: uuid }
        command: { type: [string, "null"] }
        output: { type: [string, "null"] }
        onFailure:
          oneOf:
            - { $ref: "#/components/schemas/OnFailureAction" }
            - { type: "null" }
        retryCount: { type: [integer, "null"] }
        taskOrder: { type: [integer, "null"] }
      required: [taskId]